Entry moved to _entry.py
