+++ dsadmin/tools.py
@@ -1,0 +1,428 @@
+"""Tools for creating and managing servers
+
+"""
+try:
+    from subprocess import Popen, PIPE, STDOUT
+    HASPOPEN = True
+except ImportError:
+    import popen2
+    HASPOPEN = False
+
+import sys
+import os
+import os.path
+import base64
+import urllib
+import urllib2
+import ldap
+import operator
+import select, time
+
+from dsadmin.utils import *
+from dsadmin import utils
+import logging
+logging.basicConfig(level=logging.DEBUG)
+log = logging.getLogger(__name__)
+
+
+class DSAdminTools(object):
+    """DSAdmin mix-in."""
+
+    @staticmethod
+    def cgiFake(sroot, verbose, prog, args):
+        """Run the local program prog as a CGI using the POST method."""
+        content = urllib.urlencode(args)
+        length = len(content)
+        # setup CGI environment
+        env = os.environ.copy()
+        env['REQUEST_METHOD'] = "POST"
+        env['NETSITE_ROOT'] = sroot
+        env['CONTENT_LENGTH'] = str(length)
+        progdir = os.path.dirname(prog)
+        if HASPOPEN:
+            pipe = Popen(prog, cwd=progdir, env=env,
+                         stdin=PIPE, stdout=PIPE, stderr=STDOUT)
+            child_stdin = pipe.stdin
+            child_stdout = pipe.stdout
+        else:
+            saveenv = os.environ
+            os.environ = env
+            child_stdout, child_stdin = popen2.popen2(prog)
+            os.environ = saveenv
+        child_stdin.write(content)
+        child_stdin.close()
+        for line in child_stdout:
+            if verbose:
+                sys.stdout.write(line)
+            ary = line.split(":")
+            if len(ary) > 1 and ary[0] == 'NMC_Status':
+                exitCode = ary[1].strip()
+                break
+        child_stdout.close()
+        if HASPOPEN:
+            osCode = pipe.wait()
+            print "%s returned NMC code %s and OS code %s" % (
+                prog, exitCode, osCode)
+        return exitCode
+
+    @staticmethod
+    def cgiPost(host, port, username, password, uri, verbose, secure, args=None):
+        """Post the request to the admin server.
+
+           Admin server requires authentication, so we use the auth handler classes.
+
+            NOTE: the url classes in python use the deprecated
+            base64.encodestring() function, which truncates lines,
+            causing Apache to give us a 400 Bad Request error for the
+            Authentication string.  So, we have to tell
+            base64.encodestring() not to truncate."""
+        args = args or {}
+        prefix = 'http'
+        if secure:
+            prefix = 'https'
+        hostport = host + ":" + port
+        # construct our url
+        url = '%s://%s:%s%s' % (prefix, host, port, uri)
+        # tell base64 not to truncate lines
+        savedbinsize = base64.MAXBINSIZE
+        base64.MAXBINSIZE = 256
+        # create the password manager - we don't care about the realm
+        passman = urllib2.HTTPPasswordMgrWithDefaultRealm()
+        # add our password
+        passman.add_password(None, hostport, username, password)
+        # create the auth handler
+        authhandler = urllib2.HTTPBasicAuthHandler(passman)
+        # create our url opener that handles basic auth
+        opener = urllib2.build_opener(authhandler)
+        # make admin server think we are the console
+        opener.addheaders = [('User-Agent', 'Fedora-Console/1.0')]
+        if verbose:
+            print "requesting url", url
+            sys.stdout.flush()
+        exitCode = 1
+        try:
+            req = opener.open(url, urllib.urlencode(args))
+            for line in req:
+                if verbose:
+                    print line
+                ary = line.split(":")
+                if len(ary) > 1 and ary[0] == 'NMC_Status':
+                    exitCode = ary[1].strip()
+                    break
+            req.close()
+#         except IOError, e:
+#             print e
+#             print e.code
+#             print e.headers
+#             raise
+        finally:
+            # restore binsize
+            base64.MAXBINSIZE = savedbinsize
+        return exitCode
+
+
+    def serverCmd(self, cmd, verbose, timeout=120):
+        """NOTE: this tries to open the log!
+        """
+        instanceDir = os.path.join(self.sroot, "slapd-" + self.inst)
+
+        errLog = instanceDir + '/logs/errors'
+        if hasattr(self, 'errlog'):
+            errLog = self.errlog
+        done = False
+        started = True
+        lastLine = ""
+        cmd = cmd.lower()
+        fullCmd = instanceDir + "/" + cmd + "-slapd"
+        if cmd == 'start':
+            cmdPat = 'slapd started.'
+        else:
+            cmdPat = 'slapd stopped.'
+
+        if "USE_GDB" in os.environ or "USE_VALGRIND" in os.environ:
+            timeout = timeout * 3
+        timeout += int(time.time())
+        if cmd == 'stop':
+            log.warn("unbinding before stop")
+            self.unbind()
+        
+        log.info("Setup error log")
+        logfp = open(errLog, 'r')
+        logfp.seek(0, os.SEEK_END)  # seek to end
+        pos = logfp.tell()  # get current position
+        logfp.seek(pos, os.SEEK_SET)  # reset the EOF flag
+        
+        log.warn("Running command: %r" % fullCmd)
+        rc = os.system(fullCmd)
+        while not done and int(time.time()) < timeout:
+            line = logfp.readline()
+            while not done and line:
+                lastLine = line
+                if verbose:
+                    log.debug("current line: %r" % line.strip())
+                if line.find(cmdPat) >= 0:
+                    started += 1
+                    if started == 2:
+                        done = True
+                elif line.find("Initialization Failed") >= 0:
+                    # sometimes the server fails to start - try again
+                    rc = os.system(fullCmd)
+                elif line.find("exiting.") >= 0:
+                    # possible transient condition - try again
+                    rc = os.system(fullCmd)
+                pos = logfp.tell()
+                line = logfp.readline()
+            if line.find("PR_Bind") >= 0:
+                # server port conflicts with another one, just report and punt
+                log.debug("last line: %r" % lastLine.strip())
+                log.warn("This server cannot be started until the other server on this port is shutdown")
+                done = True
+            if not done:
+                time.sleep(2)
+                logfp.seek(pos, 0)
+        logfp.close()
+        if started < 2:
+            now = int(time.time())
+            if now > timeout:
+                log.warn(
+                    "Probable timeout: timeout=%d now=%d" % (timeout, now))
+
+            log.error("Error: could not %s server %s %s: %d" % (
+                      cmd, self.sroot, self.inst, rc))
+            return 1
+        else:
+            log.info("%s was successful for %s %s" % (
+                     cmd, self.sroot, self.inst))
+            if cmd == 'start':
+                self.__localinit__()
+        return 0
+
+
+
+    def stop(self, verbose=False, timeout=0):
+        """Stop server or raise."""
+        if not self.isLocal and hasattr(self, 'asport'):
+            log.info("stopping remote server ", self)
+            self.unbind()
+            log.info("closed remote server ", self)
+            cgiargs = {}
+            rc = DSAdminTools.cgiPost(self.host, self.asport, self.cfgdsuser,
+                                      self.cfgdspwd,
+                                      "/slapd-%s/Tasks/Operation/stop" % self.inst,
+                                      verbose, cgiargs)
+            log.info("stopped remote server %s rc = %d" % (self, rc))
+            return rc
+        else:
+            return self.serverCmd('stop', verbose, timeout)
+
+    def start(self, verbose=False, timeout=0):
+        if not self.isLocal and hasattr(self, 'asport'):
+            log.debug("starting remote server %s " % self)
+            cgiargs = {}
+            rc = DSAdminTools.cgiPost(self.host, self.asport, self.cfgdsuser,
+                                      self.cfgdspwd,
+                                      "/slapd-%s/Tasks/Operation/start" % self.inst,
+                                      verbose, cgiargs)
+            log.debug("connecting remote server %s" % self)
+            if not rc:
+                self.__localinit__()
+            log.info("started remote server %s rc = %d" % (self, rc))
+            return rc
+        else:
+            log.debug("Starting server %r" % self)
+            return self.serverCmd('start', verbose, timeout)
+
+    @staticmethod
+    def runInfProg(prog, content, verbose):
+        """run a program that takes an .inf style file on stdin"""
+        cmd = [prog]
+        if verbose:
+            cmd.append('-ddd')
+        else:
+            cmd.extend(['-l', '/dev/null'])
+        cmd.extend(['-s', '-f', '-'])
+        print "running: %s " % cmd
+        if HASPOPEN:
+            pipe = Popen(cmd, stdin=PIPE, stdout=PIPE, stderr=STDOUT)
+            child_stdin = pipe.stdin
+            child_stdout = pipe.stdout
+        else:
+            pipe = popen2.Popen4(cmd)
+            child_stdin = pipe.tochild
+            child_stdout = pipe.fromchild
+        child_stdin.write(content)
+        child_stdin.close()
+        while not pipe.poll():
+            (rr, wr, xr) = select.select([child_stdout], [], [], 1.0)
+            if rr and len(rr) > 0:
+                line = rr[0].readline()
+                if not line:
+                    break
+                if verbose:
+                    sys.stdout.write(line)
+            elif verbose:
+                print "timed out waiting to read from", cmd
+        child_stdout.close()
+        exitCode = pipe.wait()
+        if verbose:
+            print "%s returned exit code %s" % (prog, exitCode)
+        return exitCode
+
+    @staticmethod
+    def createInstance(args):
+        """Create a new instance of directory server.  First, determine the hostname to use.  By
+        default, the server will be created on the localhost.  Also figure out if the given
+        hostname is the local host or not."""
+        verbose = args.get('verbose', 0)
+        isLocal = getnewhost(args)
+
+        # old style or new style?
+        sroot = args.get('sroot', os.environ.get('SERVER_ROOT', None))
+        if sroot and 'sroot' not in args:
+            args['sroot'] = sroot
+        # new style - prefix or FHS?
+        prefix = args.get('prefix', os.environ.get('PREFIX', None))
+        if 'prefix' not in args:
+            args['prefix'] = (prefix or '')
+        args['new_style'] = not sroot
+
+        # do we have ds only or ds+admin?
+        if 'no_admin' not in args:
+            sbindir = get_sbin_dir(sroot, prefix)
+            if os.path.isfile(sbindir + '/setup-ds-admin.pl'):
+                args['have_admin'] = True
+
+        if 'have_admin' not in args:
+            args['have_admin'] = False
+
+        # get default values from adm.conf
+        if args['new_style'] and args['have_admin']:
+            admconf = LDIFConn(
+                args['prefix'] + "/etc/dirsrv/admin-serv/adm.conf")
+            args['admconf'] = admconf.get('')
+
+        # next, get the configuration ds host and port
+        if args['have_admin']:
+            args['cfgdshost'], args[
+                'cfgdsport'], cfgdn = getcfgdsinfo(args)
+        if args['have_admin']:
+            cfgconn = getcfgdsuserdn(cfgdn, args)
+        # next, get the server root if not given
+        if not args['new_style']:
+            getserverroot(cfgconn, isLocal, args)
+        # next, get the admin domain
+        if args['have_admin']:
+            getadmindomain(isLocal, args)
+        # next, get the admin server port and any other information - close the cfgconn
+        if args['have_admin']:
+            asport, secure = getadminport(cfgconn, cfgdn, args)
+        # next, get the server user id
+        getserveruid(args)
+        # fixup and verify other args
+        if 'newport' not in args:
+            args['newport'] = '389'
+        if 'newrootdn' not in args:
+            args['newrootdn'] = 'cn=directory manager'
+        if 'newsuffix' not in args:
+            args['newsuffix'] = getdefaultsuffix(args['newhost'])
+        if not isLocal or 'cfgdshost' in args:
+            if 'admin_domain' not in args:
+                args['admin_domain'] = getdomainname(args['newhost'])
+            if isLocal and 'cfgdspwd' not in args:
+                args['cfgdspwd'] = "dummy"
+            if isLocal and 'cfgdshost' not in args:
+                args['cfgdshost'] = args['newhost']
+            if isLocal and 'cfgdsport' not in args:
+                args['cfgdsport'] = 55555
+        missing = False
+        for param in ('newhost', 'newport', 'newrootdn', 'newrootpw', 'newinst', 'newsuffix'):
+            if param not in args:
+                print "missing required argument", param
+                missing = True
+        if missing:
+            raise InvalidArgumentError("missing required arguments")
+
+        # try to connect with the given parameters
+        try:
+            newconn = DSAdmin(args['newhost'], args['newport'],
+                              args['newrootdn'], args['newrootpw'])
+            newconn.isLocal = isLocal
+            if args['have_admin']:
+                newconn.asport = asport
+                newconn.cfgdsuser = args['cfgdsuser']
+                newconn.cfgdspwd = args['cfgdspwd']
+            print "Warning: server at %s:%s already exists, returning connection to it" % \
+                  (args['newhost'], args['newport'])
+            return newconn
+        except ldap.SERVER_DOWN:
+            pass  # not running - create new one
+
+        if not isLocal or 'cfgdshost' in args:
+            for param in ('cfgdshost', 'cfgdsport', 'cfgdsuser', 'cfgdspwd', 'admin_domain'):
+                if param not in args:
+                    print "missing required argument", param
+                    missing = True
+        if not isLocal and not asport:
+            print "missing required argument admin server port"
+            missing = True
+        if missing:
+            raise InvalidArgumentError("missing required arguments")
+
+        # construct a hash table with our CGI arguments - used with cgiPost
+        # and cgiFake
+        cgiargs = {
+            'servname': args['newhost'],
+            'servport': args['newport'],
+            'rootdn': args['newrootdn'],
+            'rootpw': args['newrootpw'],
+            'servid': args['newinst'],
+            'suffix': args['newsuffix'],
+            'servuser': args['newuserid'],
+            'start_server': 1
+        }
+        if 'cfgdshost' in args:
+            cgiargs['cfg_sspt_uid'] = args['cfgdsuser']
+            cgiargs['cfg_sspt_uid_pw'] = args['cfgdspwd']
+            cgiargs['ldap_url'] = "ldap://%s:%d/%s" % (
+                args['cfgdshost'], args['cfgdsport'], cfgdn)
+            cgiargs['admin_domain'] = args['admin_domain']
+
+        if not isLocal:
+            DSAdmin.cgiPost(args['newhost'], asport, args['cfgdsuser'],
+                            args['cfgdspwd'], "/slapd/Tasks/Operation/Create", verbose,
+                            secure, cgiargs)
+        elif not args['new_style']:
+            prog = args['sroot'] + "/bin/slapd/admin/bin/ds_create"
+            if not os.access(prog, os.X_OK):
+                prog = args['sroot'] + "/bin/slapd/admin/bin/ds_newinst"
+            DSAdmin.cgiFake(args['sroot'], verbose, prog, cgiargs)
+        else:
+            prog = ''
+            if args['have_admin']:
+                prog = get_sbin_dir(sroot, prefix) + "/setup-ds-admin.pl"
+            else:
+                prog = get_sbin_dir(sroot, prefix) + "/setup-ds.pl"
+            content = DSAdmin.formatInfData(args)
+            DSAdmin.runInfProg(prog, content, verbose)
+
+        newconn = DSAdmin(args['newhost'], args['newport'],
+                          args['newrootdn'], args['newrootpw'])
+        newconn.isLocal = isLocal
+        if args['have_admin']:
+            newconn.asport = asport
+            newconn.cfgdsuser = args['cfgdsuser']
+            newconn.cfgdspwd = args['cfgdspwd']
+        return newconn
+
+    @staticmethod
+    def createAndSetupReplica(createArgs, repArgs):
+        # pass this sub two dicts - the first one is a dict suitable to create
+        # a new instance - see createInstance for more details
+        # the second is a dict suitable for replicaSetupAll - see replicaSetupAll
+        conn = DSAdmin.createInstance(createArgs)
+        if not conn:
+            print "Error: could not create server", createArgs
+            return 0
+
+        conn.replicaSetupAll(repArgs)
+        return conn
